+++ internal/directstream/manager.go
@@ -51,6 +51,8 @@ type (
 		// ---------- Playback State ---------- //
 
 		currentStream         mo.Option[Stream] // The current stream being played
+		currentPlaybackID     string
+		currentPlaybackClient string
 		preparingClientID     string
 		preparationCanceled   bool
 		preparationCancelFunc func()